moved from callbacks to promises. writing to .env.local moved to promptUser from makeAPI.
