+++ bin/auth0setup/create.js
@@ -1,10 +1,8 @@
 var fs = require('fs')
-var request = require('request');
+var rp = require('request-promise-native')
 
 var create = function() {
 	
-	var makeClient = require('./makeClient.js')
-
 	// get a token
 	var options = { method: 'POST',
 		url: 'https://' + process.env.DOMAIN + '/oauth/token',
@@ -18,14 +16,10 @@ var create = function() {
 		json: true 
 	}
 
-	request(options, function (error, response, body) {
-		if(error) throw error; // TODO: add human readable error message for bad url/dns error.
-	
-		process.env.ACCESS_TOKEN = body.access_token		
-		// use the management api to create a new client
-		makeClient()
-
-	});
+	var result = rp(options)
+	return result.then(function(body){
+		process.env.ACCESS_TOKEN = body.access_token
+	})
 	
 }
