+++ bin/auth0setup/promptUser.js
@@ -1,34 +1,57 @@
 var fs = require('fs')
 var dotenv = require('dotenv')
-var prompt = require('prompt')
 dotenv.config({path: '.sample-env'})
+var prompt = require('prompt')
 var create = require('./create.js')
+var makeClient = require('./makeClient.js')
+var makeAPI = require('./makeAPI.js')
+
+var envLocation = '.env.local'
 
 
 var getUserInfo = function() {
 	prompt.get(['DOMAIN', 'EXPLORER_CLIENT_ID', 'EXPLORER_CLIENT_SECRET'], function (err, result) {
 		if (err) { return onErr(err); }
-
-			
-		// var envFile = fs.readFileSync('.sample-env', 'utf8')
-		// var envJson = dotenv.parse(envFile)
 		
 		process.env.DOMAIN = result.DOMAIN
 		process.env.EXPLORER_CLIENT_ID = result.EXPLORER_CLIENT_ID
-		process.env.EXPLORER_CLIENT_SECRET = result.EXPLORER_CLIENT_SECRET		
-		
-		// envJson.DOMAIN = result.DOMAIN
-		// envJson.EXPLORER_CLIENT_ID = result.EXPLORER_CLIENT_ID
-		// envJson.EXPLORER_CLIENT_SECRET = result.EXPLORER_CLIENT_SECRET
+		process.env.EXPLORER_CLIENT_SECRET = result.EXPLORER_CLIENT_SECRET
 
-		// var envString = ''
-		
-		// for(var k in envJson) {
-			// envString = envString + k + '=' + envJson[k] + '\n'
-		// }
-		// fs.writeFileSync('.env.local', envString)
-		
 		create()
+		.then(makeClient)
+		.then(makeAPI)
+		.then(function(){
+			var envJson = {
+				REACT_APP_AUTH0_CLIENT_ID: process.env.REACT_APP_AUTH0_CLIENT_ID,
+				REACT_APP_AUTH0_DOMAIN: process.env.REACT_APP_AUTH0_DOMAIN,
+				REACT_APP_AUTH0_API_IDENTIFIER: process.env.REACT_APP_AUTH0_API_IDENTIFIER,
+				REACT_APP_GRAPHCOOL_API: ''
+			}
+			var envString = ''
+			for(var k in envJson) {
+				envString = envString + k + '=' + envJson[k] + '\n'
+			}
+			fs.writeFileSync(envLocation, envString)			
+		})
+		
+		// var createResult = create()
+		// var makeClientResult = createResult.then(makeClient)
+		// var makeAPIResult = makeClientResult.then(makeAPI)
+		// makeAPIResult.then(function(){
+			// // Take created environment variables and save to .env.local
+			// var envJson = {
+				// REACT_APP_AUTH0_CLIENT_ID: process.env.REACT_APP_AUTH0_CLIENT_ID,
+				// REACT_APP_AUTH0_DOMAIN: process.env.REACT_APP_AUTH0_DOMAIN,
+				// REACT_APP_AUTH0_API_IDENTIFIER: process.env.REACT_APP_AUTH0_API_IDENTIFIER,
+				// REACT_APP_GRAPHCOOL_API: ''
+			// }
+			// var envString = ''
+			// for(var k in envJson) {
+				// envString = envString + k + '=' + envJson[k] + '\n'
+			// }
+
+			// fs.writeFileSync(envLocation, envString)			
+		// })
 		
 	});	
 }
@@ -36,12 +59,12 @@ var getUserInfo = function() {
 // If .env.local exists, ask the user whether they want to overwrite the file or not.
 if(fs.existsSync(process.cwd() + '/.env.local')) {
 
-	var promptarray = [{'description': 'A .env.local file was found. Do you want to overwrite it? [yes/no]', name: 'overwrite'}]
+	var promptarray = [{'description': 'A .env.local file was found. Do you want to overwrite it? [y/n]', name: 'overwrite'}]
 	prompt.start()
 	prompt.get(promptarray, function(err,result){
 		if(err) { return onErr(err) }
 		
-		if(result.overwrite) {
+		if(result.overwrite == 'y') {
 			getUserInfo()
 		}
