+++ package.json
@@ -30,11 +30,12 @@
   },
   "homepage": "https://members.codefordenver.org",
   "devDependencies": {
+    "auth0-js": "^8.11.2",
+    "dotenv": "^4.0.0",
     "dotenv-cli": "^1.4.0",
     "eslint": "^4.13.1",
     "graphcool": "^0.9.2",
-		"auth0-js": "^8.11.2",
-		"dotenv": "^4.0.0",
-		"request": "^2.83.0"
+    "request": "^2.83.0",
+    "request-promise-native": "^1.0.5"
   }
 }
